# [Lenovo X1] gui-vm: Enable GPU acceleration