@vilvo 2024-08-31
Tested to work as expected in `gui-vm`.
Testing-team note: do not expect GPU acceleration in `chrome-vm` or other VMs - also `chrome://gpu` reports SW rendering - as expected. The GPU is isolated to the `gui-vm` only. @dmitry-erin - please note this also in the `ghaf-ctrl-panel` integration - if we want MESA support, the control panel app must run inside the `gui-vm`