@leivos-unikie 2024-09-02
Test results on Lenovo-X1
- PR defined tests ok 
- All apps launch
- ci-test-automation passed, performance ok
- Noticed a disturbing flash of logs while doing log out and log in. This does not happen with the current ghaf mainline build. Maybe change in kernel version has induced it? See below:

Log out test 1
![log_out_1](https://github.com/user-attachments/assets/d1cbf550-23d8-412d-8825-6c9c0c05b48f)


Log out test 2
![log_out_2](https://github.com/user-attachments/assets/d51a4c09-e7da-4994-96f2-dd00af3a1c80)


Log in test
![log_in_2](https://github.com/user-attachments/assets/2bf43b28-f424-4d6b-ba34-d1c4814fbf74)