@vunnyso 2024-09-02
> Test results on Lenovo-X1
> 
> * PR defined tests ok
> * All apps launch
> * ci-test-automation passed, performance ok
> * Noticed a disturbing flash of logs while doing log out and log in. This does not happen with the current ghaf mainline build. Maybe change in kernel version has induced it? See below:
> 
> Log out test 1 ![log_out_1](https://private-user-images.githubusercontent.com/129752236/363625252-d1cbf550-23d8-412d-8825-6c9c0c05b48f.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjUyNzYzMzAsIm5iZiI6MTcyNTI3NjAzMCwicGF0aCI6Ii8xMjk3NTIyMzYvMzYzNjI1MjUyLWQxY2JmNTUwLTIzZDgtNDEyZC04ODI1LTZjOWMwYzA1YjQ4Zi5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQwOTAyJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MDkwMlQxMTIwMzBaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT00MWUzY2NjNzdhOGU1MjZhYmIxNGJiOWY1OTA5NzA3NzNkYjMzMGYyNWM5NWY2NWQ1NDJmYzhiMmM2Njg0OGQ4JlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCZhY3Rvcl9pZD0wJmtleV9pZD0wJnJlcG9faWQ9MCJ9.ytsAVOx7nQu5F012gPf6KIeqtbiE_iHD_njWhRlx0kk)
> 
> Log out test 2 ![log_out_2](https://private-user-images.githubusercontent.com/129752236/363625696-d51a4c09-e7da-4994-96f2-dd00af3a1c80.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjUyNzYzMzAsIm5iZiI6MTcyNTI3NjAzMCwicGF0aCI6Ii8xMjk3NTIyMzYvMzYzNjI1Njk2LWQ1MWE0YzA5LWU3ZGEtNDk5NC05NmYyLWRkMDBhZjNhMWM4MC5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQwOTAyJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MDkwMlQxMTIwMzBaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT03YjA5ZTViZjEzYTFjZmNmOTEwOGRlNTYwMDVlMmZjN2NmYjdlODdjNmJkOGYwNWQwYzVlZWFmOWQ5YjZiYTg1JlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCZhY3Rvcl9pZD0wJmtleV9pZD0wJnJlcG9faWQ9MCJ9.5Yv1XTtc9BJRD-sRIaDgHN3hTPJu9sOPZCkxD0RMMxA)
> 
> Log in test ![log_in_2](https://private-user-images.githubusercontent.com/129752236/363626791-2906ec37-db18-40a5-b60e-012341bc179d.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjUyNzYzMzAsIm5iZiI6MTcyNTI3NjAzMCwicGF0aCI6Ii8xMjk3NTIyMzYvMzYzNjI2NzkxLTI5MDZlYzM3LWRiMTgtNDBhNS1iNjBlLTAxMjM0MWJjMTc5ZC5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQwOTAyJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MDkwMlQxMTIwMzBaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT1iZjFlYTZmOTU3NzIxNWZmNmVlN2QwNjZiODA5ZjE0YTNmNGIyZDhhY2UzN2IwYWM3MjJjZTVkYzMzM2I4NzQzJlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCZhY3Rvcl9pZD0wJmtleV9pZD0wJnJlcG9faWQ9MCJ9.zd5ElV0-zlj7pLmh-AyMqGi3TKaPvmS5VknlMjFlLpk)

Thanks @leivos-unikie for the findings, I have tried with mainline kernel (6.6.43) as well, I have noticed same errors with GPU acceleration, I have fixed CURSOR and "Fontconfig" issues, rest I think are warnings which can be ignored as of now unless resulting in any regression. 

**Side Note:** If possible please check GPU acceleration on Gen12 or Gen13 if available.