@vilvo 2024-09-02
> Test results on Lenovo-X1
> 
> * PR defined tests ok
> * All apps launch
> * ci-test-automation passed, performance ok

Thanks for testing. This is good.

> * Noticed a disturbing flash of logs while doing log out and log in. This does not happen with the current ghaf mainline build. Maybe change in kernel version has induced it? See below:

Reading logs, touchpad driver errors are not critical but could/should be worked outside the scope of this PR. 
I read three distinct issues on those:
~~1. setting optional cursor size for the graphical session compositor (`export XCURSOR_SIZE`)~~ - FIXED
~~2. no writable cache for fonts~~ - FIXED
3. touchpad kernel driver bug - it can be also be related to GPU accelerated cursor setting that is changed in this PR. Seems right button is lost (at least sometimes) and some ghost touchpad input events happen. At least I found no issues with using the touchpad but if there are fixed for those, sure they could be worked separately.