@leivos-unikie 2024-09-03
Seems better. Now the error log in login is at least shorter.

Lot out
![log_out](https://github.com/user-attachments/assets/ff9e4ff2-f792-47e3-80ac-2c9bba964c4b)

Log in
![log_in](https://github.com/user-attachments/assets/78875a9a-f110-4d18-b084-f5b7c4465b4e)